release: add auth with jwt

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './enums/task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,12 +14,18 @@ export class TasksService {
     private taskRepository : TaskRepository,
   ) {}
 
-  async getTasks(filterDTO: GetTasksFilterDTO): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDTO);
+  async getTasks(
+    filterDTO: GetTasksFilterDTO,
+    user: User,
+    ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDTO, user);
   }
 
-  async getTaskById(id: number): Promise<Task>{
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+    ): Promise<Task>{
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found){
       throw new NotFoundException(`Task with ID "${id}" not found`)
     }
@@ -26,20 +33,30 @@ export class TasksService {
     return found;
   }
 
-  async createTask(createTaskDTO:CreateTaskDTO): Promise<Task> {
-    return this.taskRepository.createTaskDTO(createTaskDTO);
+  async createTask(
+    createTaskDTO:CreateTaskDTO,
+    user:User,
+    ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDTO, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const found = await this.taskRepository.delete(id);
+  async deleteTask(
+    id: number,
+    user: User,
+    ): Promise<void> {
+    const found = await this.taskRepository.delete({ id, userId: user.id });
 
     if(found.affected === 0){
       throw new NotFoundException(`Task with ID "${id}" not found`) 
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number, 
+    status: TaskStatus,
+    user: User,
+    ): Promise<Task>{
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
